Feat: Add isActualParentGroup flag in useGraphTransform
This change is a precursor to Task 4 (EntityNode Enhancements)
 as per Section 9, Phase 1, Step 2 of the PRD.

It adds an `isActualParentGroup` boolean flag to the `data` object
of nodes that serve as parents in a compound node structure.
This flag will be used by EntityNode.jsx to apply distinct styling
to parent group nodes.

Also removed the unused layoutType from imports and params as Dagre is now the sole layout.